\name{VRclust-package}
\alias{VRclust-package}
\alias{VRclust}
\docType{package}
\title{VRclust: An optimal clustering algorithm with varying risks of false assignments across clusters}
\description{
There are three components to this package. These are: i) Construct and fit the finite mixture model when data are multivariate normal or multinomial (\code{vrclust.fit}). ii) Construct the optimal clustering rule for controlling the false assignment rate (\code{FAR.control}). iii) Obtain the estimated confusion matrix (\code{vrclust.est}).
}
